refactor(Step-2 피드백 반영): SplitUtils 클래스 StringUtils로 통합
