--- src/main/java/calculator/util/SplitUtils.java
@@ -1,12 +1,0 @@
-package calculator.util;
-
-import java.util.Arrays;
-
-public class SplitUtils {
-
-    public static int[] splitToInt(String text, String separator) throws NumberFormatException {
-        return Arrays.stream(text.split(separator))
-                    .mapToInt(Integer::parseInt)
-                    .toArray();
-    }
-}
