+++ src/main/java/calculator/util/StringUtils.java
@@ -1,5 +1,7 @@
 package calculator.util;
 
+import java.util.Arrays;
+
 public class StringUtils {
 
     private StringUtils() {
@@ -8,5 +10,11 @@ public class StringUtils {
 
     public static boolean isEmptyString(String string) {
         return string == null || string.isEmpty();
+    }
+
+    public static int[] splitToInt(String text, String separator) throws NumberFormatException {
+        return Arrays.stream(text.split(separator))
+                .mapToInt(Integer::parseInt)
+                .toArray();
     }
 }
